+++ Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+*empty*
+
+## [1.4.0] - 2022-05-27
 ### Added
 - Add ability to print out location of config directory
 - Create view for empty sections
@@ -33,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/badjr13/laydown
 <!-- Obtained by going to last commit before version bump and `Browse Files` -->
+[1.4.0]: https://github.com/badjr13/laydown/tree/0bb487815550b9182398508c22b8134b30844724
 [1.3.0]: https://github.com/badjr13/laydown/tree/f6b23dcdd5b86796831e8e7f81282bb8341aad91
 [1.2.0]: https://github.com/badjr13/laydown/tree/5a10fe65baac320d6a557a66af6372f690c04118
 [1.1.0]: https://github.com/badjr13/laydown/tree/5b8c4a82a2362e0ed6a902e8166bb570f6dab403